pyproject.toml
[project]
name = "rag-platform"
version = "0.1.0"
description = "Shared RAG platform with multi-tenant isolation and MCP server"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "google-genai>=1.0.0",
    "numpy>=1.26.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "google-auth>=2.20.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.14.0",
]
mcp = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
